Fix types and itemMemory indexing
Accuracy is close to 100% with pure Python code and was about 75% with the C implementation.
I've pinned it down to the way _itemMemory is passed and indexed by the C code.
I suspect the tensor simply wasn't contiguous in memory. For now, it's bypassed by means
of Python code performing the indexing into the correct item.